requirements.txt
uvicorn
fastapi
psycopg2
pyjwt
pydantic[email]
passlib